# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      description: Admin only. Returns all registered users.
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/AppUser"
                required:
                  - users
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      description: Admin only. Creates a new user with the default password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/AppUser"
                required:
                  - user
        "400":
          description: Missing email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}/reset-password:
    put:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset a user's password
      description: Admin only. Resets the user's password to the default and sets mustChangePassword.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  email:
                    type: string
                required:
                  - success
                  - email
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}/role:
    put:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's role
      description: Admin only. Sets the user's role to "admin" or "user".
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetRoleRequest"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/activity:
    get:
      operationId: getUserActivity
      tags: [users]
      summary: Get login activity for all users
      description: Admin only. Returns per-user login sessions grouped by day.
      responses:
        "200":
          description: User activity data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserActivityResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      description: Admin only. Deletes a user account. Cannot delete your own account.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
        "400":
          description: Cannot delete own account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AppUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        role:
          type: string
        mustChangePassword:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - role
        - mustChangePassword
        - createdAt
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, user]
      required:
        - email
    UserSessionEntry:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        loginAt:
          type: string
        lastActivityAt:
          type: ["string", "null"]
          description: ISO-8601. Last heartbeat received; null for legacy rows. Used to determine idle vs active.
        logoutAt:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
      required:
        - id
        - userId
        - email
        - loginAt
    UserActivityDay:
      type: object
      properties:
        date:
          type: string
          description: YYYY-MM-DD
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/UserSessionEntry"
      required:
        - date
        - sessions
    UserActivityResponse:
      type: object
      properties:
        days:
          type: array
          items:
            $ref: "#/components/schemas/UserActivityDay"
        totalSessions:
          type: integer
      required:
        - days
        - totalSessions
...
    SetRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum: [admin, user]
      required:
        - role
...